Add a folder entity Small correction
