+++ src/lib.rs
@@ -17,6 +17,5 @@ pub mod world_dispatcher;
 pub mod cubes_to_draw;
 pub mod world_serializer;
 pub mod network;
-pub mod entity;
 pub mod math;
-pub mod humanoid;
+pub mod entity;
